Attempting to print, and learning
Printing the value with every half/complete callback doesn't work presumably because the callback and the UART DMA peripheral are trying to read the same memory at the same time.

It got slightly better when I set Clock Prescaler to Synchronous clock mode divided by 4, more time

It seems when I try to write using the complete callback, it gets to the half callback faster than UART prints it, so the second half of the string is not sent.

Very cool!